Refactoring Pomodoro-related files and folders
1) Move the Brainwave folder into the pomodoro folder

2) Separate logic from timer.js file: timerButton.js, progressBar.js

3) Link javascript and css files to related files, respectively
