+++ pomodoro.html
@@ -6,33 +6,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="src/reset.css" />
     <link rel="stylesheet" href="src/variables.css" />
-    <link
-      rel="stylesheet"
-      href="src/components/pomodoro/timer/progress-bar/progress-bar.css"
-    />
     <link rel="stylesheet" href="src/components/pomodoro/pomodoro.css" />
-    <link rel="stylesheet" href="src/components/brainwave/brainwave.css" />
     <link
       href="https://fonts.googleapis.com/icon?family=Material+Icons"
       rel="stylesheet"
     />
-
-    <script
-      src="https://code.jquery.com/jquery-3.6.4.min.js"
-      integrity="sha256-oP6HI9z1XaZNBrJURtCoUT5SUnxFr8s3BzRl+cbzUq8="
-      crossorigin="anonymous"
-    ></script>
-
-    <script defer src="src/components/brainwave/brainwave.js"></script>
-    <script
-      defer
-      src="src/components/pomodoro/timer/progress-bar/progress-bar.js"
-    ></script>
-    <script
-      defer
-      type="module"
-      src="src/components/pomodoro/pomodoro.js"
-    ></script>
+    <script type="module" src="src/components/pomodoro/pomodoro.js"></script>
     <title>Pomodoro</title>
   </head>
   <body>
