+++ src/components/pomodoro/timer/timer.css
@@ -15,42 +15,3 @@
   text-shadow: 2px 2px 2px rgba(0, 0, 0, 0.18);
   margin-top: 35px;
 }
-
-.timer__button {
-  width: 125px;
-  height: 50px;
-  margin: 0 20px;
-  margin-top: 110px;
-  font-size: 32px;
-  color: white;
-  border-radius: 25px;
-  box-shadow: var(--button-shadow);
-  -webkit-box-shadow: var(--button-shadow);
-  -moz-box-shadow: va;
-}
-
-.timer__button:active:not(.disabled) {
-  box-shadow: -8px 8px 5px -3px rgba(0, 0, 0, 0.33) inset;
-  -webkit-box-shadow: -8px 8px 5px -3px rgba(0, 0, 0, 0.33) inset;
-  -moz-box-shadow: -8px 8px 5px -3px rgba(0, 0, 0, 0.33) inset;
-}
-
-.timer__button:first-of-type {
-  background-color: var(--third-theme-color);
-}
-
-.timer__button:last-of-type {
-  background-color: #92d050;
-}
-
-.timer__button.disabled {
-  background-color: #9e9e9e !important;
-  box-shadow: none;
-  -webkit-box-shadow: none;
-  -moz-box-shadow: none;
-  cursor: default;
-}
-
-.timer__text.rest {
-  -webkit-text-stroke: 1px rgba(0, 0, 0, 0.25);
-}
